interface: merge models/interface methods into actions/interface
This is to have all the defined interfaces in one place
and no duplicates
